Update(data type): Update uint to uint64 data type

+++ internal/database/seeder/access_seed.go
@@ -27,10 +27,10 @@ func seed_Access() {
 		log.Println(err)
 	}
 	if countRole == 0 {
-		service.CreateRole("admin", []uint{})
-		service.CreateRole("crew", []uint{1, 2})
+		service.CreateRole("admin", []uint64{})
+		service.CreateRole("crew", []uint64{1, 2})
 
 		// assign role to client
-		service.UpdateClient(1, []uint{1, 2})
+		service.UpdateClient(1, []uint64{1, 2})
 	}
 }
