+++ internal/domain/models/Category.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Category struct {
-	ID        uint      `json:"id" gorm:"primaryKey;autoIncrement:true;unique"`
+	ID        uint64      `json:"id" gorm:"primaryKey;autoIncrement:true;unique"`
 	Name      string    `json:"name" gorm:"not null" validate:"required"`
 	CreatedAt time.Time `gorm:"autoCreateTime"`
 }
